requirements.txt
# -------------------------
# FastAPI Backend
# -------------------------
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9

# -------------------------
# Machine Learning
# -------------------------
tensorflow==2.15.0
numpy==1.26.4

# -------------------------
# Audio Processing
# -------------------------
librosa==0.10.1
soundfile==0.12.1

# -------------------------
# Image Processing
# -------------------------
Pillow==10.2.0
matplotlib==3.8.3

# -------------------------
# Utilities
# -------------------------
scipy==1.12.0
reportlab==4.1.0
